requirements.txt
pytest>=7.0
requests>=2.28
PyYAML
jsonschema
pytest-xdist
pytest-html
tenacity
behave
python-dotenv
allure-pytest